main.go refactoring: define cmd action as a separate func.
This splits out the scraper code into a separate file next to main.go,
and wraps the action code into a separate function that is used as
default action of the urfave app. Setting a default action prevents
the help being printed out unnecessarily.
